refactor(app.py): Phase 5b — extract settings + profile + admin blueprints
Move the remaining 14 route handlers out of app.py into three new
Flask blueprints under web/:

- web/settings.py (settings_bp, 527 lines) — 5 routes: /settings,
  /settings/config, /api/validate-keys, /api/providers/reorder,
  /api/job-sources/<key>/toggle
- web/profile.py (profile_bp, 649 lines) — 4 routes: /profile,
  /profile/import-pdf, /profile/import-pdf/status/<job_id>,
  /api/apply-prefilter-suggestions
- web/admin.py (admin_bp, 310 lines) — 5 routes (/admin,
  /admin/clear-db, /admin/logs, /admin/logs/<file>/download,
  /admin/schedule-state) plus _LOG_FILENAME_RE and SCHEDULE_*
  constants

app.py drops from 1475 → 336 lines (−1139). All three blueprints
register with url_prefix="" in create_app(), and endpoint="settings" /
endpoint="profile" overrides preserve url_for("settings") at
app.py:447 and url_for("profile") at app.py:944 unchanged. Route
count stays at 29 — zero URL changes, zero template changes needed.

Test patch migrations (no dual-copy — per Phase 4 lesson):
- test_admin_logs.py, test_credential_source_bugs.py,
  test_pdf_async.py, test_profile.py, test_profile_import.py,
  test_reorder.py, test_settings.py, test_settings_save.py,
  test_toggle_source.py, test_validate_search_config.py — patches
  migrated from app.X to web.admin.X / web.profile.X / web.settings.X
  (Python's "patch where used, not where defined" rule).

Verification (full CI parity locally):
- pytest: 2045 passed, 1 skipped, 0 failed (same as post-5a baseline)
- ruff check .: clean
- python -c "import app": 29 routes register (unchanged)
- shellcheck: skipped (not installed locally; CI will run it)

refs #330
