+++ tests/test_settings_save.py
@@ -51,6 +51,7 @@ import pytest
 sys.path.insert(0, os.path.join(os.path.dirname(__file__), ".."))
 
 import app as app_module
+import web.settings as _settings_module
 from app import app as flask_app
 from credentials import save_providers
 
@@ -65,6 +66,7 @@ def tmp_providers_path(tmp_path, monkeypatch):
     """Point _PROVIDERS_PATH at a temp file for full isolation."""
     path = str(tmp_path / "providers.json")
     monkeypatch.setattr(app_module, "_PROVIDERS_PATH", path)
+    monkeypatch.setattr(_settings_module, "_PROVIDERS_PATH", path)
     return path
 
 
@@ -73,6 +75,7 @@ def tmp_keys_path(tmp_path, monkeypatch):
     """Point _KEYS_PATH at a temp file so legacy migration never triggers."""
     path = str(tmp_path / "keys.json")
     monkeypatch.setattr(app_module, "_KEYS_PATH", path)
+    monkeypatch.setattr(_settings_module, "_KEYS_PATH", path)
     return path
 
 
@@ -81,6 +84,7 @@ def tmp_config_path(tmp_path, monkeypatch):
     """Point _CONFIG_PATH at a temp file so config reads are isolated."""
     path = str(tmp_path / "config.json")
     monkeypatch.setattr(app_module, "_CONFIG_PATH", path)
+    monkeypatch.setattr(_settings_module, "_CONFIG_PATH", path)
     return path
